requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
audioop-lts==0.2.1
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.4.26
charset-normalizer==3.4.2
deprecation==2.1.0
discord.py==2.5.2
frozenlist==1.6.0
gotrue==2.12.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
multidict==6.4.4
outcome==1.3.0.post0
packaging==25.0
pluggy==1.6.0
postgrest==1.0.2
propcache==0.3.1
pydantic==1.10.14
PyJWT==2.10.1
PySocks==1.7.1
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
realtime==2.4.3
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.1
supafunc==0.9.4
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
validators==0.35.0
vt-py==0.20.0
websocket-client==1.8.0
websockets==14.2
wsproto==1.2.0
yarl==1.20.0
spotipy==2.25.1
yt-dlp==2025.5.22